@@ -1,3 +1,3 @@
 module github.com/fabianpallares/json
 
-go 1.16
+go 1.18
